Finish moving existing processing to event processors.
