+++ aquarius/events/events_monitor.py
@@ -4,38 +4,28 @@
 #
 import json
 import logging
-import lzma as Lzma
 import os
 import time
 from datetime import datetime
 from json import JSONDecodeError
 from threading import Thread
 
-import ecies
 import elasticsearch
-import eth_keys
 import requests
 from eth_account import Account
-from eth_utils import add_0x_prefix, remove_0x_prefix
 from ocean_lib.config_provider import ConfigProvider
 from oceandb_driver_interface import OceanDb
-from plecos.plecos import is_valid_dict_remote, list_errors_dict_remote
 
-from aquarius.app.auth_util import compare_eth_addresses, sanitize_addresses
-from aquarius.app.util import (
-    DATETIME_FORMAT,
-    format_timestamp,
-    get_bool_env_value,
-    get_metadata_from_services,
-    init_new_ddo,
-    list_errors,
-    validate_data,
-)
+from aquarius.app.auth_util import sanitize_addresses
+from aquarius.app.util import get_bool_env_value
 from aquarius.block_utils import BlockProcessingClass
 from aquarius.events.constants import EVENT_METADATA_CREATED, EVENT_METADATA_UPDATED
 from aquarius.events.metadata_updater import MetadataUpdater
-from aquarius.events.util import get_datatoken_info, get_metadata_contract
-from aquarius.events.processors import MetadataCreatedProcessor
+from aquarius.events.util import get_metadata_contract
+from aquarius.events.processors import (
+    MetadataCreatedProcessor,
+    MetadataUpdatedProcessor,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -273,14 +263,19 @@ class EventsMonitor(BlockProcessingClass):
         debug_log(
             f"Metadata monitor >>>> from_block:{from_block}, current_block:{current_block} <<<<"
         )
+
+        processor_args = [
+            self._oceandb,
+            self._web3,
+            self._ecies_account,
+            self._allowed_publishers,
+        ]
+
         for event in self.get_event_logs(
             EVENT_METADATA_CREATED, from_block, current_block
         ):
             try:
-                event_processor = MetadataCreatedProcessor(
-                    event, self._oceandb, self._web3, self._ecies_account
-                )
-                event_processor.setAllowedPublishers(self._allowed_publishers)
+                event_processor = MetadataCreatedProcessor(*([event] + processor_args))
                 event_processor.process()
             except Exception as e:
                 logger.error(
@@ -291,7 +286,8 @@ class EventsMonitor(BlockProcessingClass):
             EVENT_METADATA_UPDATED, from_block, current_block
         ):
             try:
-                self.processUpdateDDO(event)
+                event_processor = MetadataUpdatedProcessor(*([event] + processor_args))
+                event_processor.process()
             except Exception as e:
                 logger.error(
                     f"Error processing update metadata event: {e}\n" f"event={event}"
@@ -346,259 +342,3 @@ class EventsMonitor(BlockProcessingClass):
             logger.error(
                 f"get_event_logs ({event_name}, {from_block}, {to_block}) failed: {e}."
             )
-
-    def is_publisher_allowed(self, publisher_address):
-        logger.debug(f"checking allowed publishers: {publisher_address}")
-        if not self._allowed_publishers:
-            return True
-
-        publisher_address = self._web3.toChecksumAddress(publisher_address)
-        return publisher_address in self._allowed_publishers
-
-    def processNewDDO(self, event):
-        (
-            did,
-            block,
-            txid,
-            contract_address,
-            sender_address,
-            flags,
-            rawddo,
-            timestamp,
-        ) = self.get_event_data(event)
-        logger.info(
-            f"Process new DDO, did from event log:{did}, sender:{sender_address}"
-        )
-        if not self.is_publisher_allowed(sender_address):
-            logger.warning(f"Sender {sender_address} is not in ALLOWED_PUBLISHERS.")
-            return
-
-        try:
-            self._oceandb.read(did)
-            logger.warning(f"{did} is already registered")
-            return
-        except Exception:
-            pass
-
-        logger.info(f"Start processing {EVENT_METADATA_CREATED} event: did={did}")
-        debug_log(
-            f"block {block}, contract: {contract_address}, Sender: {sender_address} , txid: {txid}"
-        )
-
-        logger.debug(f"decoding with did {did} and flags {flags}")
-        data = self.decode_ddo(rawddo, flags)
-        if data is None:
-            logger.warning(f"Could not decode ddo using flags {flags}")
-            return
-
-        msg, _ = validate_data(data, f"event {EVENT_METADATA_CREATED}")
-        if msg:
-            logger.warning(msg)
-            return
-
-        _record = init_new_ddo(data, timestamp)
-        # this will be used when updating the doo
-        _record["event"] = dict()
-        _record["event"]["txid"] = txid
-        _record["event"]["blockNo"] = block
-        _record["event"]["from"] = sender_address
-        _record["event"]["contract"] = contract_address
-
-        _record["price"] = {
-            "datatoken": 0.0,
-            "ocean": 0.0,
-            "value": 0.0,
-            "type": "",
-            "exchange_id": "",
-            "address": "",
-            "pools": [],
-            "isConsumable": "",
-        }
-        dt_address = _record.get("dataToken")
-        assert dt_address == add_0x_prefix(did[len("did:op:") :])
-        if dt_address:
-            _record["dataTokenInfo"] = get_datatoken_info(dt_address)
-
-        if not is_valid_dict_remote(get_metadata_from_services(_record["service"])):
-            errors = list_errors(
-                list_errors_dict_remote, get_metadata_from_services(_record["service"])
-            )
-            logger.error(f"New ddo has validation errors: {errors}")
-            return False
-
-        _record["isInPurgatory"] = "false"
-
-        try:
-            record_str = json.dumps(_record)
-            self._oceandb.write(record_str, did)
-            _record = json.loads(record_str)
-            name = _record["service"][0]["attributes"]["main"]["name"]
-            debug_log(f"DDO saved: did={did}, name={name}, publisher={sender_address}")
-            logger.info(
-                f"Done processing {EVENT_METADATA_CREATED} event: did={did}. DDO SAVED TO DB"
-            )
-            return True
-        except (KeyError, Exception) as err:
-            logger.error(
-                f"encountered an error while saving the asset data to OceanDB: {str(err)}"
-            )
-            return False
-
-    def processUpdateDDO(self, event):
-        (
-            did,
-            block,
-            txid,
-            contract_address,
-            sender_address,
-            flags,
-            rawddo,
-            timestamp,
-        ) = self.get_event_data(event)
-        debug_log(f"Process update DDO, did from event log:{did}")
-        try:
-            asset = self._oceandb.read(did)
-        except Exception:
-            # TODO: check if this asset was deleted/hidden due to some violation issues
-            # if so, don't add it again
-            logger.warning(f"{did} is not registered, will add it as a new DDO.")
-            self.processNewDDO(event)
-            return
-
-        debug_log(
-            f"block {block}, contract: {contract_address}, Sender: {sender_address} , txid: {txid}"
-        )
-
-        # do not update if we have the same txid
-        ddo_txid = asset["event"]["txid"]
-        if txid == ddo_txid:
-            logger.warning(
-                f'asset has the same txid, no need to update: event-txid={txid} <> asset-event-txid={asset["event"]["txid"]}'
-            )
-            return
-
-        # check block
-        ddo_block = asset["event"]["blockNo"]
-        if int(block) <= int(ddo_block):
-            logger.warning(
-                f"asset was updated later (block: {ddo_block}) vs transaction block: {block}"
-            )
-            return
-
-        # check owner
-        if not compare_eth_addresses(
-            asset["publicKey"][0]["owner"], sender_address, logger
-        ):
-            logger.warning("Transaction sender must mach ddo owner")
-            return
-
-        debug_log(f"decoding with did {did} and flags {flags}")
-        data = self.decode_ddo(rawddo, flags)
-        if data is None:
-            logger.warning("Cound not decode ddo")
-            return
-
-        msg, _ = validate_data(data, "event update")
-        if msg:
-            logger.error(msg)
-            return
-
-        _record = init_new_ddo(data, timestamp)
-        # make sure that we do not alter created flag
-        _record["created"] = asset["created"]
-        # but we update 'updated'
-        _record["updated"] = format_timestamp(
-            datetime.fromtimestamp(timestamp).strftime(DATETIME_FORMAT)
-        )
-        _record["event"] = dict()
-        _record["event"]["txid"] = txid
-        _record["event"]["blockNo"] = block
-        _record["event"]["from"] = sender_address
-        _record["event"]["contract"] = contract_address
-
-        if not is_valid_dict_remote(get_metadata_from_services(_record["service"])):
-            errors = list_errors(
-                list_errors_dict_remote, get_metadata_from_services(_record["service"])
-            )
-            logger.error(f"ddo update has validation errors: {errors}")
-            return
-
-        _record["price"] = asset.get("price", {})
-        dt_address = _record.get("dataToken")
-        assert dt_address == add_0x_prefix(did[len("did:op:") :])
-        if dt_address:
-            _record["dataTokenInfo"] = get_datatoken_info(dt_address)
-
-        _record["isInPurgatory"] = asset.get("isInPurgatory", "false")
-
-        try:
-            self._oceandb.update(json.dumps(_record), did)
-            logger.info(f"updated DDO saved to db successfully (did={did}).")
-            return True
-        except (KeyError, Exception) as err:
-            logger.error(
-                f"encountered an error while updating the asset data to OceanDB: {str(err)}"
-            )
-            return
-
-    def get_event_data(self, event):
-        # TODO: move in processors
-        tx_id = event.transactionHash.hex()
-        sender = event.args.get("createdBy", event.args.get("updatedBy"))
-        blockInfo = self._web3.eth.getBlock(event.blockNumber)
-        timestamp = blockInfo["timestamp"]
-        return (
-            f"did:op:{remove_0x_prefix(event.args.dataToken)}",
-            event.blockNumber,
-            tx_id,
-            event.address,
-            sender,
-            event.args.get("flags", None),
-            event.args.get("data", None),
-            timestamp,
-        )
-
-    def decode_ddo(self, rawddo, flags):
-        # TODO move in processors
-        debug_log(f"flags: {flags}")
-        # debug_log(f'Before unpack rawddo:{rawddo}')
-        if len(flags) < 1:
-            debug_log("Set check_flags to 0!")
-            check_flags = 0
-        else:
-            check_flags = flags[0]
-        if self._only_encrypted_ddo and (not check_flags & 2):
-            logger.error("This aquarius can cache only encrypted ddos")
-            return None
-        # always start with MSB -> LSB
-        debug_log(f"checkflags: {check_flags}")
-        # bit 2:  check if ddo is ecies encrypted
-        if check_flags & 2:
-            try:
-                rawddo = self.ecies_decrypt(rawddo)
-                logger.debug(f"Decrypted to {rawddo}")
-            except (KeyError, Exception) as err:
-                logger.error(f"Failed to decrypt: {str(err)}")
-
-        # bit 1:  check if ddo is lzma compressed
-        if check_flags & 1:
-            try:
-                rawddo = Lzma.decompress(rawddo)
-                logger.debug(f"Decompressed to {rawddo}")
-            except (KeyError, Exception) as err:
-                logger.error(f"Failed to decompress: {str(err)}")
-
-        logger.debug(f"After unpack rawddo:{rawddo}")
-        try:
-            ddo = json.loads(rawddo)
-            return ddo
-        except (KeyError, Exception) as err:
-            logger.error(f"encountered an error while decoding the ddo: {str(err)}")
-            return None
-
-    def ecies_decrypt(self, rawddo):
-        # TODO: move in processors
-        if self._ecies_account is not None:
-            key = eth_keys.KeyAPI.PrivateKey(self._ecies_account.privateKey)
-            rawddo = ecies.decrypt(key.to_hex(), rawddo)
-        return rawddo
