+++ aquarius/events/processors.py
@@ -3,16 +3,20 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 from abc import ABC
+from datetime import datetime
 from eth_utils import add_0x_prefix, remove_0x_prefix
 import json
 import logging
 
 from plecos.plecos import is_valid_dict_remote, list_errors_dict_remote
+from aquarius.app.auth_util import compare_eth_addresses
 from aquarius.app.util import (
-    init_new_ddo,
-    validate_data,
+    DATETIME_FORMAT,
+    format_timestamp,
     get_metadata_from_services,
+    init_new_ddo,
     list_errors,
+    validate_data,
 )
 from aquarius.events.constants import EVENT_METADATA_CREATED
 from aquarius.events.util import get_datatoken_info
@@ -23,7 +27,7 @@ debug_log = logger.debug
 
 
 class EventProcessor(ABC):
-    def __init__(self, event, oceandb, web3, ecies_account):
+    def __init__(self, event, oceandb, web3, ecies_account, allowed_publishers):
         self.event = event
         self.did = f"did:op:{remove_0x_prefix(self.event.args.dataToken)}"
         self.block = event.blockNumber
@@ -38,15 +42,13 @@ class EventProcessor(ABC):
         self._oceandb = oceandb
         self._web3 = web3
         self.decryptor = Decryptor(ecies_account)
+        self.allowed_publishers = allowed_publishers
 
         blockInfo = self._web3.eth.getBlock(self.event.blockNumber)
         self.timestamp = blockInfo["timestamp"]
 
 
 class MetadataCreatedProcessor(EventProcessor):
-    def setAllowedPublishers(self, allowed_publishers):
-        self.allowed_publishers = allowed_publishers
-
     def is_publisher_allowed(self, publisher_address):
         logger.debug(f"checking allowed publishers: {publisher_address}")
         if not self.allowed_publishers:
@@ -141,3 +143,106 @@ class MetadataCreatedProcessor(EventProcessor):
                 f"encountered an error while saving the asset data to OceanDB: {str(err)}"
             )
             return False
+
+
+class MetadataUpdatedProcessor(EventProcessor):
+    def make_record(self, data, asset):
+        _record = init_new_ddo(data, self.timestamp)
+        # make sure that we do not alter created flag
+        _record["created"] = asset["created"]
+        # but we update 'updated'
+        _record["updated"] = format_timestamp(
+            datetime.fromtimestamp(self.timestamp).strftime(DATETIME_FORMAT)
+        )
+        _record["event"] = {
+            "txid": self.txid,
+            "blockNo": self.block,
+            "from": self.sender_address,
+            "contract": self.contract_address,
+        }
+
+        if not is_valid_dict_remote(get_metadata_from_services(_record["service"])):
+            errors = list_errors(
+                list_errors_dict_remote, get_metadata_from_services(_record["service"])
+            )
+            logger.error(f"ddo update has validation errors: {errors}")
+            return
+
+        _record["price"] = asset.get("price", {})
+        dt_address = _record.get("dataToken")
+        assert dt_address == add_0x_prefix(self.did[len("did:op:") :])
+        if dt_address:
+            _record["dataTokenInfo"] = get_datatoken_info(dt_address)
+
+        _record["isInPurgatory"] = asset.get("isInPurgatory", "false")
+
+        return _record
+
+    def process(self):
+        did, sender_address = self.did, self.sender_address
+        debug_log(f"Process update DDO, did from event log:{did}")
+        try:
+            asset = self._oceandb.read(did)
+        except Exception:
+            # TODO: check if this asset was deleted/hidden due to some violation issues
+            # if so, don't add it again
+            logger.warning(f"{did} is not registered, will add it as a new DDO.")
+            event_processor = MetadataCreatedProcessor(
+                self.event,
+                self._oceandb,
+                self._web3,
+                self._ecies_account,
+                self.allowed_publishers,
+            )
+            event_processor.process()
+            return
+
+        debug_log(
+            f"block {self.block}, contract: {self.contract_address}, Sender: {sender_address} , txid: {self.txid}"
+        )
+
+        # do not update if we have the same txid
+        ddo_txid = asset["event"]["txid"]
+        if self.txid == ddo_txid:
+            logger.warning(
+                f'asset has the same txid, no need to update: event-txid={self.txid} <> asset-event-txid={asset["event"]["txid"]}'
+            )
+            return
+
+        # check block
+        ddo_block = asset["event"]["blockNo"]
+        if int(self.block) <= int(ddo_block):
+            logger.warning(
+                f"asset was updated later (block: {ddo_block}) vs transaction block: {self.block}"
+            )
+            return
+
+        # check owner
+        if not compare_eth_addresses(
+            asset["publicKey"][0]["owner"], sender_address, logger
+        ):
+            logger.warning("Transaction sender must mach ddo owner")
+            return
+
+        debug_log(f"decoding with did {did} and flags {self.flags}")
+        data = self.decryptor.decode_ddo(self.rawddo, self.flags)
+        if data is None:
+            logger.warning("Cound not decode ddo")
+            return
+
+        msg, _ = validate_data(data, "event update")
+        if msg:
+            logger.error(msg)
+            return
+
+        _record = self.make_record(data, asset)
+
+        try:
+            self._oceandb.update(json.dumps(_record), did)
+            logger.info(f"updated DDO saved to db successfully (did={did}).")
+            return True
+        except (KeyError, Exception) as err:
+            logger.error(
+                f"encountered an error while updating the asset data to OceanDB: {str(err)}"
+            )
+            return
